[MagneticField] Move public headers to interface directory
