+++ Configuration/Skimming/interface/LeptonSkimming.h
@@ -58,7 +58,6 @@
 #include "FWCore/Common/interface/TriggerNames.h"
 #include "DataFormats/HepMCCandidate/interface/GenParticle.h"
 #include "MagneticField/Engine/interface/MagneticField.h"
-#include "MagneticField/ParametrizedEngine/src/OAEParametrizedMagneticField.h"
 #include "TrackingTools/TransientTrack/interface/TransientTrack.h"
 #include "TrackingTools/IPTools/interface/IPTools.h"
 #include "RecoVertex/KalmanVertexFit/interface/KalmanVertexFitter.h"
